Split up `abs::syntax` module.
